[theme] Simplification of the API
