+++ docs/src/pages/customization/DarkTheme.js
@@ -2,13 +2,12 @@
 
 import React from 'react';
 import { MuiThemeProvider, createMuiTheme } from 'material-ui/styles';
-import createPalette from 'material-ui/styles/palette';
 import WithTheme from './WithTheme';
 
 const theme = createMuiTheme({
-  palette: createPalette({
+  palette: {
     type: 'dark', // Switching the dark mode on is a single property value change.
-  }),
+  },
 });
 
 function DarkTheme() {
