+++ docs/src/pages/customization/Palette.js
@@ -2,21 +2,20 @@
 
 import React from 'react';
 import { MuiThemeProvider, createMuiTheme } from 'material-ui/styles';
-import createPalette from 'material-ui/styles/palette';
 import purple from 'material-ui/colors/purple';
 import green from 'material-ui/colors/green';
 import red from 'material-ui/colors/red';
 import Button from 'material-ui/Button';
 
 const theme = createMuiTheme({
-  palette: createPalette({
+  palette: {
     primary: purple, // Purple and green play nicely together.
     secondary: {
       ...green,
       A400: '#00e677',
     },
     error: red,
-  }),
+  },
 });
 
 function Palette() {
