+++ docs/src/pages/customization/TypographyTheme.js
@@ -1,33 +1,24 @@
 // @flow weak
 
 import React from 'react';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import createTypography from 'material-ui/styles/typography';
+import { MuiThemeProvider, createMuiTheme } from 'material-ui/styles';
 import Typography from 'material-ui/Typography';
 import Button from 'material-ui/Button';
 
 function theme(outerTheme) {
-  const typography = createTypography(outerTheme.palette, {
-    // System font
-    fontFamily:
-      '-apple-system,system-ui,BlinkMacSystemFont,' +
-      '"Segoe UI",Roboto,"Helvetica Neue",Arial,sans-serif',
-  });
-
-  return {
-    ...outerTheme,
+  return createMuiTheme({
     typography: {
-      ...typography,
+      fontFamily:
+        '-apple-system,system-ui,BlinkMacSystemFont,' +
+        '"Segoe UI",Roboto,"Helvetica Neue",Arial,sans-serif',
       body1: {
-        ...typography.body1,
-        fontWeight: typography.fontWeightMedium,
+        fontWeight: outerTheme.typography.fontWeightMedium,
       },
       button: {
-        ...typography.button,
         fontStyle: 'italic',
       },
     },
-  };
+  });
 }
 
 function TypographyTheme() {
