+++ examples/create-react-app/src/styles/createContext.js
@@ -3,16 +3,15 @@
 import { create } from 'jss';
 import preset from 'jss-preset-default';
 import { SheetsRegistry } from 'react-jss';
-import createPalette from 'material-ui/styles/palette';
-import createMuiTheme from 'material-ui/styles/theme';
+import { createMuiTheme } from 'material-ui/styles';
 import { purple, green } from 'material-ui/colors';
 import createGenerateClassName from 'material-ui/styles/createGenerateClassName';
 
 const theme = createMuiTheme({
-  palette: createPalette({
+  palette: {
     primary: purple,
-    accent: green,
-  }),
+    secondary: green,
+  },
 });
 
 // Configure JSS
