+++ src/index.js
@@ -39,15 +39,7 @@ export { default as Popover } from './Popover';
 export { CircularProgress, LinearProgress } from './Progress';
 export { default as Radio, RadioGroup } from './Radio';
 export { default as Snackbar, SnackbarContent } from './Snackbar';
-export {
-  MuiThemeProvider,
-  withStyles,
-  withTheme,
-  createMuiTheme,
-  createTypography,
-  createBreakpoints,
-  createPalette,
-} from './styles';
+export { MuiThemeProvider, withStyles, withTheme, createMuiTheme } from './styles';
 
 // eslint-disable-next-line import/first
 import * as colors from './colors';
