+++ src/styles/index.js
@@ -3,7 +3,4 @@
 export { default as MuiThemeProvider } from './MuiThemeProvider';
 export { default as withStyles } from './withStyles';
 export { default as withTheme } from './withTheme';
-export { default as createMuiTheme } from './theme';
-export { default as createTypography } from './typography';
-export { default as createBreakpoints } from './breakpoints';
-export { default as createPalette } from './palette';
+export { default as createMuiTheme } from './createMuiTheme';
